Move the responsibility of merging ElasticSearch and SCSB responses into the AvailabilityResolver
