+++ .env.example
@@ -1,0 +1,2 @@
+SCSB_URL=https://get.from.coworker:9093
+SCSB_API_KEY=getmefromacowoker
