+++ lib/resources.js
@@ -1,8 +1,7 @@
-var ResourceResultsSerializer = require('./jsonld_serializers.js').ResourceResultsSerializer
 var ResourceSerializer = require('./jsonld_serializers.js').ResourceSerializer
 var AggregationsSerializer = require('./jsonld_serializers.js').AggregationsSerializer
 var AggregationSerializer = require('./jsonld_serializers.js').AggregationSerializer
-var request = require('request')
+var AvailabilityResolver = require('./availability_resolver.js')
 var util = require('../lib/util')
 var config = require('config')
 
@@ -144,81 +143,8 @@ module.exports = function (app) {
       index: RESOURCES_INDEX,
       body: body
     }).then((resp) => {
-      // This can be a Set. To ensure uniqueness.
-      var barcodesAndAvailability = {}
-      var barcodes = []
-      for (let hit of resp.hits.hits) {
-        for (let item of hit._source.items) {
-          // despite its singular name item.identifier is an Array of barcodes
-          if (item.identifier) {
-            for (let identifier of item.identifier) {
-              let barcode_array = identifier.split(':')
-              if (barcode_array.length === 3 && barcode_array[1] === 'barcode') { barcodes.push(barcode_array[barcode_array.length - 1]) }
-            }
-          }
-        }
-      }
-      return getItemsAvailabilityStatus(barcodes)
-        .then((itemsStatus) => {
-          itemsStatus.forEach((statusEntry) => {
-            barcodesAndAvailability[statusEntry.itemBarcode] = statusEntry.itemAvailabilityStatus
-          })
-          return barcodesAndAvailability
-        })
-        .then((barcodesAndAvailability) => {
-          fixItemAvailabilityInResponse(resp, barcodesAndAvailability)
-          return ResourceResultsSerializer.serialize(resp)
-        })
-        .catch((error) => {
-          app.logger.error('Error occurred while setting availability - ', error)
-          return Promise.reject(error)
-        })
-    })
-  }
-
-  var fixItemAvailabilityInResponse = (resp, barcodesAndAvailability) => {
-    for (let hit of resp.hits.hits) {
-      hit._source.items.map((item) => {
-        if (item.identifier) {
-          for (let identifier of item.identifier) {
-            let barcode_array = identifier.split(':')
-            if (barcode_array.length === 3 && barcode_array[1] === 'barcode') {
-              if (barcodesAndAvailability[barcode_array[barcode_array.length - 1]] !== "Item Barcode doesn't exist in SCSB database.") {
-                // can an item have multiple status
-                if (barcodesAndAvailability[barcode_array[barcode_array.length - 1]] === 'Available') {
-                  item.status[0].id = config.get('itemAvailability.available.id')
-                  item.status[0].label = config.get('itemAvailability.available.label')
-                } else {
-                  item.status[0].id = config.get('itemAvailability.temporarilyUnavailable.id')
-                  item.status[0].label = config.get('itemAvailability.temporarilyUnavailable.label')
-                }
-              }
-            }
-          }
-        }
-        return item
-      })
-    }
-  }
-
-  var getItemsAvailabilityStatus = (listBarcodes) => {
-    return new Promise((resolve, reject) => {
-      var bodyToSend = {}
-      bodyToSend = {barcodes: listBarcodes}
-      var options = {
-        url: config.get('recap.itemAvailabilityStatus.url'),
-        headers: {
-          'Accept': 'application/json',
-          'api_key': config.get('recap.itemAvailabilityStatus.api-key'),
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(bodyToSend)
-      }
-      request.post(options, (error, response, body) => {
-        if (error) { reject(error) } else if (response && response.statusCode === 200) {
-          resolve(JSON.parse(body))
-        }
-      })
+      let availabilityResolver = new AvailabilityResolver(resp)
+      return availabilityResolver.responseWithUpdatedAvailability()
     })
   }
